client gfx refactor (4): finish wrapping lobby
