+++ alvr/client_openxr/src/lobby.rs
@@ -2,9 +2,10 @@ use crate::{
     graphics::{self, CompositionLayerBuilder},
     interaction, XrContext,
 };
-use alvr_client_core::RenderViewInput;
+use alvr_client_core::{GraphicsContext, LobbyRenderer, RenderViewInput};
 use alvr_common::glam::UVec2;
 use openxr as xr;
+use std::rc::Rc;
 
 // todo: add interaction?
 pub struct Lobby {
@@ -13,10 +14,16 @@ pub struct Lobby {
     swapchains: [xr::Swapchain<xr::OpenGlEs>; 2],
     view_resolution: UVec2,
     reference_space_type: xr::ReferenceSpaceType,
+    renderer: LobbyRenderer,
 }
 
 impl Lobby {
-    pub fn new(xr_ctx: &XrContext, view_resolution: UVec2) -> Self {
+    pub fn new(
+        xr_ctx: &XrContext,
+        gfx_ctx: Rc<GraphicsContext>,
+        view_resolution: UVec2,
+        initial_hud_message: &str,
+    ) -> Self {
         let reference_space_type = if xr_ctx.instance.exts().ext_local_floor.is_some() {
             xr::ReferenceSpaceType::LOCAL_FLOOR_EXT
         } else {
@@ -33,7 +40,8 @@ impl Lobby {
             graphics::create_swapchain(&xr_ctx.session, view_resolution, None, false),
         ];
 
-        alvr_client_core::opengl::initialize_lobby(
+        let renderer = LobbyRenderer::new(
+            gfx_ctx,
             view_resolution,
             [
                 swapchains[0]
@@ -50,6 +58,7 @@ impl Lobby {
                     .collect(),
             ],
             false, // TODO: correct lobby sRGB for some headsets
+            initial_hud_message,
         );
 
         Self {
@@ -58,12 +67,17 @@ impl Lobby {
             swapchains,
             view_resolution,
             reference_space_type,
+            renderer,
         }
     }
 
     pub fn update_reference_space(&mut self) {
         self.reference_space =
             interaction::get_reference_space(&self.xr_session, self.reference_space_type);
+    }
+
+    pub fn update_hud_message(&mut self, message: &str) {
+        self.renderer.update_hud_message(message);
     }
 
     pub fn render(&mut self, predicted_display_time: xr::Time) -> CompositionLayerBuilder {
@@ -92,7 +106,7 @@ impl Lobby {
             .wait_image(xr::Duration::INFINITE)
             .unwrap();
 
-        alvr_client_core::opengl::render_lobby([
+        self.renderer.render([
             RenderViewInput {
                 pose: crate::from_xr_pose(views[0].pose),
                 fov: crate::from_xr_fov(views[0].fov),
